refactor: TableService가 TableValidator를 주입 받아서 사용하도록 변경
