+++ src/main/java/kitchenpos/application/TableService.java
@@ -1,17 +1,13 @@
 package kitchenpos.application;
 
-import static kitchenpos.exception.ErrorCode.EXISTS_NOT_COMPLETION_STATUS;
 import static kitchenpos.exception.ErrorCode.NOT_EXISTS_TABLE;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.validator.TableValidator;
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.dto.response.OrderTableResponse;
 import kitchenpos.exception.KitchenposException;
-import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,14 +16,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class TableService {
     private final OrderTableRepository orderTableRepository;
-    private final OrderRepository orderRepository;
+    private final TableValidator tableValidator;
 
     public TableService(
             final OrderTableRepository orderTableRepository,
-            final OrderRepository orderRepository
+            final TableValidator tableValidator
     ) {
+        this.tableValidator = tableValidator;
         this.orderTableRepository = orderTableRepository;
-        this.orderRepository = orderRepository;
     }
 
     public OrderTableResponse create(final OrderTable orderTable) {
@@ -48,8 +44,7 @@ public class TableService {
 
     public OrderTableResponse changeEmpty(final Long orderTableId, final boolean empty) {
         final OrderTable savedOrderTable = findById(orderTableId);
-        TableValidator.validateChangeEmpty(savedOrderTable);
-        existsCookingOrMeal(orderTableId);
+        tableValidator.validateChangeEmpty(savedOrderTable);
 
         savedOrderTable.changeEmpty(empty);
 
@@ -57,24 +52,15 @@ public class TableService {
     }
 
     public OrderTableResponse changeNumberOfGuests(final Long orderTableId, final OrderTable orderTable) {
-        TableValidator.validateNumberOfGuests(orderTable.getNumberOfGuests());
+        tableValidator.validateNumberOfGuests(orderTable.getNumberOfGuests());
 
         final int numberOfGuests = orderTable.getNumberOfGuests();
 
         final OrderTable savedOrderTable = findById(orderTableId);
-        TableValidator.validateEmptyTrue(savedOrderTable);
+        tableValidator.validateEmptyTrue(savedOrderTable);
 
         savedOrderTable.setNumberOfGuests(numberOfGuests);
 
         return OrderTableResponse.of(orderTableRepository.save(savedOrderTable));
-    }
-
-    private void existsCookingOrMeal(Long orderTableId) {
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
-                orderTableId,
-                Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))
-        ) {
-            throw new KitchenposException(EXISTS_NOT_COMPLETION_STATUS);
-        }
     }
 }
